Extract old ads module
