+++ app/modules/ads/module/new-ads.js
@@ -1,0 +1,9 @@
+class NewAds {
+  static getInstance() {
+    return;
+  }
+
+  static getLoadedInstance() {
+    return;
+  }
+}
